requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.5.0
    # via django
attrs==21.4.0
    # via pytest
babel==2.9.1
    # via py-moneyed
black==22.1.0
    # via -r requirements.in
boto3==1.21.10
    # via -r requirements.in
botocore==1.24.10
    # via
    #   boto3
    #   s3transfer
brotli==1.0.9
    # via fonttools
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via weasyprint
cfgv==3.3.1
    # via pre-commit
charset-normalizer==2.0.12
    # via requests
click==8.0.4
    # via
    #   black
    #   pip-tools
    #   tryceratops
colorama==0.4.4
    # via
    #   pretty-errors
    #   rich
commonmark==0.9.1
    # via rich
coverage[toml]==6.3.2
    # via
    #   -r requirements.in
    #   django-coverage-plugin
    #   pytest-cov
crispy-tailwind==0.5.0
    # via -r requirements.in
cssselect2==0.5.0
    # via weasyprint
distlib==0.3.4
    # via virtualenv
django==4.0.3
    # via
    #   -r requirements.in
    #   django-csp
    #   django-debug-toolbar
    #   django-extensions
    #   django-money
    #   django-nplusone
    #   django-permissions-policy
    #   django-storages
    #   django-weasyprint
    #   djangorestframework
django-admin-interface==0.18.7
    # via -r requirements.in
django-colorfield==0.6.3
    # via django-admin-interface
django-coverage-plugin==2.0.2
    # via -r requirements.in
django-crispy-forms==1.14.0
    # via crispy-tailwind
django-csp==3.7
    # via -r requirements.in
django-debug-toolbar==3.2.4
    # via -r requirements.in
django-environ==0.8.1
    # via -r requirements.in
django-extensions==3.1.5
    # via -r requirements.in
django-feature-policy==4.0.0
    # via -r requirements.in
django-flat-responsive==2.0
    # via django-admin-interface
django-flat-theme==1.1.4
    # via django-admin-interface
django-money==2.1.1
    # via -r requirements.in
django-nplusone==0.0.5
    # via -r requirements.in
django-permissions-policy==4.8.0
    # via django-feature-policy
django-storages==1.12.3
    # via -r requirements.in
django-weasyprint==2.1.0
    # via -r requirements.in
djangorestframework==3.13.1
    # via -r requirements.in
djhtml==1.5.0
    # via -r requirements.in
docutils==0.18.1
    # via -r requirements.in
execnet==1.9.0
    # via pytest-xdist
factory-boy==3.2.1
    # via -r requirements.in
faker==13.3.0
    # via factory-boy
filelock==3.6.0
    # via virtualenv
flake8==4.0.1
    # via -r requirements.in
fonttools[woff]==4.29.1
    # via weasyprint
gunicorn==20.1.0
    # via -r requirements.in
html5lib==1.1
    # via weasyprint
identify==2.4.11
    # via pre-commit
idna==3.3
    # via requests
iniconfig==1.1.1
    # via pytest
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
lxml==4.8.0
    # via python-docx-1
markdown2==2.4.2
    # via -r requirements.in
mccabe==0.6.1
    # via flake8
mypy-extensions==0.4.3
    # via black
nodeenv==1.6.0
    # via pre-commit
packaging==21.3
    # via pytest
pathspec==0.9.0
    # via black
pep517==0.12.0
    # via pip-tools
pillow==9.0.1
    # via
    #   django-colorfield
    #   weasyprint
pip-tools==6.5.1
    # via -r requirements.in
platformdirs==2.5.1
    # via
    #   black
    #   virtualenv
pluggy==1.0.0
    # via pytest
pre-commit==2.17.0
    # via -r requirements.in
pretty-errors==1.2.25
    # via -r requirements.in
psycopg2-binary==2.9.3
    # via -r requirements.in
py==1.11.0
    # via
    #   pytest
    #   pytest-forked
py-moneyed==1.2
    # via django-money
pycodestyle==2.8.0
    # via flake8
pycparser==2.21
    # via cffi
pydot==1.4.2
    # via -r requirements.in
pydyf==0.1.2
    # via weasyprint
pyflakes==2.4.0
    # via flake8
pygments==2.11.2
    # via rich
pygraphviz==1.9
    # via -r requirements.in
pyparsing==3.0.7
    # via
    #   -r requirements.in
    #   packaging
    #   pydot
pypdf2==1.26.0
    # via -r requirements.in
pyphen==0.12.0
    # via weasyprint
pytest==7.0.1
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-django
    #   pytest-forked
    #   pytest-randomly
    #   pytest-xdist
pytest-cov==3.0.0
    # via -r requirements.in
pytest-django==4.5.2
    # via -r requirements.in
pytest-forked==1.4.0
    # via pytest-xdist
pytest-randomly==3.11.0
    # via -r requirements.in
pytest-xdist==2.5.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   botocore
    #   faker
python-docx-1==0.0.2
    # via -r requirements.in
pytz==2021.3
    # via
    #   babel
    #   djangorestframework
pyyaml==6.0
    # via pre-commit
requests==2.27.1
    # via -r requirements.in
rich==10.16.2
    # via tryceratops
s3transfer==0.5.2
    # via boto3
six==1.16.0
    # via
    #   django-admin-interface
    #   django-coverage-plugin
    #   html5lib
    #   python-dateutil
    #   virtualenv
sqlparse==0.4.2
    # via
    #   django
    #   django-debug-toolbar
tinycss2==1.1.1
    # via
    #   cssselect2
    #   weasyprint
toml==0.10.2
    # via
    #   pre-commit
    #   tryceratops
tomli==2.0.1
    # via
    #   black
    #   coverage
    #   pep517
    #   pytest
tryceratops==1.0.1
    # via -r requirements.in
urllib3==1.26.8
    # via
    #   botocore
    #   requests
virtualenv==20.13.2
    # via pre-commit
weasyprint==54.2
    # via django-weasyprint
webencodings==0.5.1
    # via
    #   cssselect2
    #   html5lib
    #   tinycss2
wheel==0.37.1
    # via pip-tools
zopfli==0.1.9
    # via fonttools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
